[REFACTOR] 클래스 이름 수정 및 로그 삭제
